requirements.txt
streamlit==1.32.0
folium==0.16.0
streamlit-folium==0.18.0
plotly==5.19.0
pandas==2.2.0
numpy==1.26.4
rasterio==1.3.9
scikit-learn==1.4.0
xgboost==2.0.3
shap==0.44.1
africastalking==1.2.5
requests==2.31.0
python-dotenv==1.0.0